server/requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
colorama==0.4.4
Django==3.1.7
django-colorfield==0.4.1
django-cors-headers==3.7.0
djangorestframework==3.12.2
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
python-decouple==3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wincertstore==0.2
wrapt==1.12.1
